Refactoring to break up files into more components and hooks
