+++ parking-lot/src/components/modals/ConfirmDeleteModal.js
