+++ parking-lot/src/hooks/useRedirect.js
@@ -1,0 +1,11 @@
+import { useEffect } from 'react';
+
+const useRedirect = (isLoggedIn, history) => {
+
+  useEffect(() => {
+    const path = isLoggedIn ? "/home" : "/";
+    history.push(path);
+  }, [isLoggedIn, history]);
+}
+
+export default useRedirect;
